+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
 - Configuration for asset service
 
+##### Fixed
+
+- Deprecated JS event
+
 #### [2.2.0] - 2025-11-02
 
 ##### Added